Improved performance of Day 12.
Paths are no longer explicitly stored only to count them afterwards. Instead, their count is intermediately added and created during recursive calls.
The current Path is a Set instead of a List, since we discard the Path anyways. The order is not relevant and a Set provides faster lookup times.